Fixes process list running before creation
The `subprocess.run` to run process lists was running before
`subprocess.Popen` to create process lists was created.

Fixed this by calling `subprocess.Popen.communicate() and checking
that the stderr is None, before printing to the terminal to inform
user that process list is created.

Also prints the stdout of running the SavuRecon process so that user
knows when it has failed, rather than just a generic "reconstruction
complete" even when it did not complete correctly.