Allows centre of rotation to be a single line
If the centre of rotation is set to 'autocenter' in the `savurecon.yaml`
file, then use the centre of the stack as the centre of rotation.

If the centre of rotation is set to a float, use this value for all ts.

If the centre of rotation is not a float or 'autocenter' then raise
ValueError.